pyproject.toml
[project]
name = "worldcup2026"
version = "0.1.0"
description = "MLOps pipeline to predict the FIFA World Cup 2026 champion"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "MLOps WorldCup 2026 Team" }]
license = { text = "MIT" }

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.3",
    "requests>=2.31.0",
    "pyarrow>=15.0.0",
    "mlflow>=2.11.0",
    "prefect>=2.16.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "typer>=0.9.0",
    "joblib>=1.3.2",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.3.0",
    "httpx>=0.27.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
wc2026 = "worldcup2026.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worldcup2026"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]